Menghapus personal info dan merubahnya menjadi class
